fix: colocate runtime metrics CLI tests without new pkg/cli file
Keep runtime metrics flag tests in root_run_server_test.go so pkg/cli
stays within the 15-file package limit after trimming root_run_test.go.
